this

+++ 0x00-MySQL_Advanced/101-average_weighted_score.sql
@@ -7,7 +7,7 @@ RETURNS FLOAT DETERMINISTIC
 BEGIN
     DECLARE total_weighted_score INT DEFAULT 0;
     DECLARE total_weight INT DEFAULT 0;
-	DECLARE result INT DEFAULT 0;
+	DECLARE result FLOAT DEFAULT 0;
 
     SELECT SUM(corrections.score * projects.weight)
         INTO total_weighted_score
